Project updated for Pkcs11Interop 2.0

+++ src/Pkcs11Interop.PkiUtils/ObjectImporter.cs
@@ -1,26 +1,16 @@
 /*
  *  Pkcs11Interop.PkiUtils - PKI extensions for Pkcs11Interop library
- *  Copyright (c) 2013 JWC s.r.o.
- *  Author: Jaroslav Imrich
- *  
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License version 3
- *  as published by the Free Software Foundation.
- *  
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU Affero General Public License for more details.
- *  
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program. If not, see <http://www.gnu.org/licenses/>.
- *  
- *  You can be released from the requirements of the license by purchasing
- *  a commercial license. Buying such a license is mandatory as soon as you
- *  develop commercial activities involving the Pkcs11Interop.PkiUtils software 
- *  without disclosing the source code of your own applications.
- *  
- *  For more information, please contact JWC s.r.o. at info@pkcs11interop.net
+ *  Copyright (c) 2013-2014 JWC s.r.o. <http://www.jwc.sk>
+ *  Author: Jaroslav Imrich <jimrich@jimrich.sk>
+ *
+ *  Licensing for open source projects:
+ *  Pkcs11Interop.PkiUtils is available under the terms of the GNU Affero General 
+ *  Public License version 3 as published by the Free Software Foundation.
+ *  Please see <http://www.gnu.org/licenses/agpl-3.0.html> for more details.
+ *
+ *  Licensing for other types of projects:
+ *  Pkcs11Interop.PkiUtils is available under the terms of flexible commercial license.
+ *  Please contact JWC s.r.o. at <info@pkcs11interop.net> for more details.
  */
 
 using System;
@@ -36,15 +26,15 @@ namespace Net.Pkcs11Interop.PkiUtils
     /// <summary>
     /// Utilities for object importing
     /// </summary>/summary>
-	public static class ObjectImporter
-	{
+    public static class ObjectImporter
+    {
         /// <summary>
         /// Imports the certificate into the PKCS#11 compatible device and pairs it with the corresponding private key
         /// </summary>
         /// <param name="session">Session with user logged in</param>
         /// <param name="certificate">Certificate that should be imported</param>
         /// <returns>Handle of created certificate object</returns>
-		public static ObjectHandle ImportCertificate(Session session, byte[] certificate)
+        public static ObjectHandle ImportCertificate(Session session, byte[] certificate)
         {
             // Parse certificate
             X509CertificateParser x509CertificateParser = new X509CertificateParser();
@@ -58,8 +48,8 @@ namespace Net.Pkcs11Interop.PkiUtils
 
             // Find corresponding private key
             List<ObjectAttribute> privKeySearchTemplate = new List<ObjectAttribute>();
-            privKeySearchTemplate.Add(new ObjectAttribute(CKA.CKA_CLASS, (uint)CKO.CKO_PRIVATE_KEY));
-            privKeySearchTemplate.Add(new ObjectAttribute(CKA.CKA_KEY_TYPE, (uint)CKK.CKK_RSA));
+            privKeySearchTemplate.Add(new ObjectAttribute(CKA.CKA_CLASS, CKO.CKO_PRIVATE_KEY));
+            privKeySearchTemplate.Add(new ObjectAttribute(CKA.CKA_KEY_TYPE, CKK.CKK_RSA));
             privKeySearchTemplate.Add(new ObjectAttribute(CKA.CKA_MODULUS, rsaPubKeyParams.Modulus.ToByteArrayUnsigned()));
             privKeySearchTemplate.Add(new ObjectAttribute(CKA.CKA_PUBLIC_EXPONENT, rsaPubKeyParams.Exponent.ToByteArrayUnsigned()));
 
@@ -78,12 +68,12 @@ namespace Net.Pkcs11Interop.PkiUtils
 
             // Define attributes of new certificate object
             List<ObjectAttribute> certificateAttributes = new List<ObjectAttribute>();
-            certificateAttributes.Add(new ObjectAttribute(CKA.CKA_CLASS, (uint)CKO.CKO_CERTIFICATE));
+            certificateAttributes.Add(new ObjectAttribute(CKA.CKA_CLASS, CKO.CKO_CERTIFICATE));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_TOKEN, true));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_PRIVATE, false));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_MODIFIABLE, true));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_LABEL, privKeyAttributes[0].GetValueAsString()));
-            certificateAttributes.Add(new ObjectAttribute(CKA.CKA_CERTIFICATE_TYPE, (uint)CKC.CKC_X_509));
+            certificateAttributes.Add(new ObjectAttribute(CKA.CKA_CERTIFICATE_TYPE, CKC.CKC_X_509));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_TRUSTED, false));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_SUBJECT, x509Certificate.SubjectDN.GetDerEncoded()));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_ID, privKeyAttributes[1].GetValueAsByteArray()));
@@ -126,12 +116,12 @@ namespace Net.Pkcs11Interop.PkiUtils
 
             // Define attributes of new certificate object
             List<ObjectAttribute> certificateAttributes = new List<ObjectAttribute>();
-            certificateAttributes.Add(new ObjectAttribute(CKA.CKA_CLASS, (uint)CKO.CKO_CERTIFICATE));
+            certificateAttributes.Add(new ObjectAttribute(CKA.CKA_CLASS, CKO.CKO_CERTIFICATE));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_TOKEN, true));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_PRIVATE, false));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_MODIFIABLE, true));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_LABEL, ckaLabel));
-            certificateAttributes.Add(new ObjectAttribute(CKA.CKA_CERTIFICATE_TYPE, (uint)CKC.CKC_X_509));
+            certificateAttributes.Add(new ObjectAttribute(CKA.CKA_CERTIFICATE_TYPE, CKC.CKC_X_509));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_TRUSTED, true));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_SUBJECT, x509Certificate.SubjectDN.GetDerEncoded()));
             certificateAttributes.Add(new ObjectAttribute(CKA.CKA_ID, ckaId));
@@ -142,5 +132,5 @@ namespace Net.Pkcs11Interop.PkiUtils
             // Create certificate object
             return session.CreateObject(certificateAttributes);
         }
-	}
+    }
 }
